Extract DPM PM quality client routes
