+++ quality/ci_quality_gates.md
@@ -62,12 +62,12 @@ Most recent local PR-grade evidence:
 
 1. The previous quality-baseline enforcement branch added
    `scripts/check_refactor_quality_thresholds.py` as a blocking lint-stage gate.
-2. Current enforced source-file threshold: no Python source file under `src/app` above 1,362
+2. Current enforced source-file threshold: no Python source file under `src/app` above 1,237
    script-counted lines.
 3. Current enforced function threshold: no Python function or async function above the remediated
    49-line AST span baseline.
 4. `python scripts/check_refactor_quality_thresholds.py`: passed with
-   `max_source_file_lines=1362` and `max_function_lines=49`.
+   `max_source_file_lines=1237` and `max_function_lines=49`.
 5. Feature Lane and PR Merge Gate step names now call out `Lint and Refactor Quality Thresholds`
    so the promoted gate is visible in GitHub logs.
 6. Current portfolio workspace payload mapper branch `make check` passed with 1,075
@@ -246,11 +246,18 @@ Most recent local PR-grade evidence:
     upstream warning and partial-failure behavior. Focused validation passed with ruff check,
     ruff format check, touched-module mypy, 67 service/component/boundary unit tests, and a trial
     refactor threshold gate at `max_source_file_lines=1362`.
-31. Current concrete-route registration fix expands registered gateway routers into concrete
+31. Current DPM PM operating-quality client-boundary branch moves score-run, fairness-analysis,
+    review-action, summary-invocation, and policy route methods into
+    `dpm_pm_operating_quality_client.py` while preserving `DpmClient` as the public upstream
+    client surface. Focused validation passed with ruff check, ruff format check, mypy over the
+    touched clients, 182 upstream-client, boundary, and threshold unit tests, and trial refactor
+    threshold gates proving `max_source_file_lines=1237` passes while `1236` fails. Local
+    `make check` passed with mypy over 484 source files and 1,159 unit/contract tests.
+32. Current concrete-route registration fix expands registered gateway routers into concrete
     `APIRoute` entries so route enumeration, contract tests, and Prometheus route-name middleware
     do not see FastAPI lazy included-router placeholders. Focused validation passed with ruff
     check, ruff format check, mypy over `router_registry.py`, and 15 route/contract tests.
-32. Current portfolio workspace component parser branch full local gates passed: `make check`
+33. Current portfolio workspace component parser branch full local gates passed: `make check`
     passed with ruff, format check, monetary-float guard, refactor threshold gate, workflow
     action-runtime gate, mypy over 483 source files, Workbench/OpenAPI contract smoke, and 1,158
     unit/contract tests. `make ci` passed with 207 integration tests, 1,365 combined coverage
@@ -265,7 +272,7 @@ Most recent local PR-grade evidence:
    whether explicit operation IDs should replace generated IDs.
 3. Promote import-linter contracts after false positives are classified.
 4. Continue tightening the enforced source-file threshold downward as the remaining largest clients
-   and services are split; `dpm_client.py` is now the largest file at 1,362 script-counted lines
-   and defines the current blocking ceiling.
+   and services are split; `portfolio_service.py` is now the largest file at 1,237 script-counted
+   lines and defines the current blocking ceiling.
 5. Extend static no-sensitive-observability checks beyond the new Prometheus metric-label gate to
    broader logs, trace attributes, and diagnostics fields.
